+++ TensorIO/tensorio/src/main/java/ai/doc/tensorio/TIOTFLiteModel/GpuDelegateHelper.java
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-package ai.doc.tensorio.TIOTensorflowLiteModel;
+package ai.doc.tensorio.TIOTFLiteModel;
 
 import org.tensorflow.lite.Delegate;
